moved global_config.py to configs/global_config.py
