+++ inputremapper/configs/global_config.py
