+++ tests/unit/test_injector.py
@@ -22,7 +22,6 @@
 import unittest
 from unittest import mock
 import time
-import copy
 
 import evdev
 from evdev.ecodes import (
@@ -30,7 +29,6 @@ from evdev.ecodes import (
     EV_KEY,
     EV_ABS,
     ABS_HAT0X,
-    BTN_LEFT,
     KEY_A,
     REL_X,
     REL_Y,
@@ -42,7 +40,6 @@ from evdev.ecodes import (
     ABS_Z,
     ABS_RZ,
     ABS_VOLUME,
-    KEY_B,
     KEY_C,
 )
 
@@ -61,7 +58,7 @@ from inputremapper.injection.numlock import is_numlock_on, set_numlock, ensure_n
 from inputremapper.system_mapping import system_mapping, DISABLE_CODE, DISABLE_NAME
 from inputremapper.gui.active_preset import active_preset
 from inputremapper.preset import Preset
-from inputremapper.global_config import global_config, NONE, MOUSE, WHEEL, BUTTONS
+from inputremapper.configs.global_config import global_config, NONE, MOUSE, WHEEL
 from inputremapper.key import Key
 from inputremapper.injection.macros.parse import parse
 from inputremapper.injection.context import Context
